refactor: move write_stream to output of impl
The write_stream takes a reference to self which make this very awkward
to do when writing to a stream within the impl. By moving this outside
of the impl so it does not take a reference self each iteration because
the tokio spawn must have a static lifetime.

This makes the function driver very similar to the actor pattern,
whereby the driving action is handled outside of the main structure.
Subjects: