feat: add `write_stream_for_duration` and `handle_futures`
`write_stream_for_duration` is a generic helper function which takes the
behaviour of writing to a stream for a particular duration and pairs it
with a predicate to end the continous loop - this loop is what makes up
the "for a duration" functionality.

`handle_futures` encapsulates the duplicate code blocks that were for
reading from tokio tasks and updating stats etc. As this is repeated in
many places, it is easy to pull it into a function.